requirements.txt
requests==2.13.0
python-dateutil==2.6.0

falcon==1.1.0

gunicorn==19.7.1